% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rutils.R
\name{sysopen}
\alias{sysopen}
\title{Open fx}
\usage{
sysopen(path_to_file)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
sysopen()
}
\keyword{cats}
